Refactor pricing components to own folder
