+++ src/pricing/components/PricingRow.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export const PricingRow = ({
+  children,
+  mobileView = false,
+  marginBottom = false,
+}) => (
+  <div
+    className={`row text-nowrap ${mobileView ? 'mobile-pricing' : 'desktop-pricing'} ${
+      marginBottom ? 'mb-4' : ''
+    }`}
+  >
+    {children}
+  </div>
+);
