+++ src/pricing/components/index.js
@@ -1,0 +1,3 @@
+export * from './PricingCol';
+export * from './PricingColChildren';
+export * from './PricingRow';
